% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adaBoost}
\alias{adaBoost}
\title{An AdaBoost classification}
\usage{
adaBoost(
  dataset,
  labels,
  iterations = 100L,
  tolerance = 2e-10,
  perceptronIter = 400L
)
}
\arguments{
\item{dataset}{A matrix of explanatory variables, i.e. \dQuote{features}}

\item{labels}{A vector of the dependent variable as integer values, i.e. \dQuote{labels}}

\item{iterations}{An integer value for the number of iterations}

\item{tolerance}{A double with the desired tolerance}

\item{perceptronIter}{An integer value for the number of a iterations for the weak learner}
}
\value{
A list object
}
\description{
Run AdaBoost using a simple Perceptron model as the weak learner
}
\examples{
data(iris)
X <- t(as.matrix(iris[,1:4]))
y <- as.integer(iris[,5]) - 1
adaBoost(X, y)
}
