% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{datasetExample}
\alias{datasetExample}
\title{Simple example of loading categorical data via 'mlpack'}
\usage{
datasetExample()
}
\value{
Nothing is returned, the function is invoked for its side effect.
}
\description{
Simple example of loading categorical data via 'mlpack'
}
