% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decisionTree}
\alias{decisionTree}
\title{Run a decisionTree classification}
\usage{
decisionTree(dataset, labels, pct = 0.3, min_leaf_size = 10L,
  minimum_gain_split = 1e-07, maximum_depth = 0L)
}
\arguments{
\item{dataset}{A matrix of explanatory variables, i.e. \dQuote{features}}

\item{labels}{A vector of the dependent variable as integer values, i.e. \dQuote{labels}}

\item{pct}{A numeric value for the percentage of data to be retained for the test set}

\item{min_leaf_size}{An integer value with the minimum number of elements per leaf}

\item{minimum_gain_split}{A double with the gain needed to further split the tree}

\item{maximum_depth}{An integer with the maximum tree depth, default zero means unlimited}
}
\value{
A list object
}
\description{
Run decisionTree classifier
}
\examples{
data(iris)
X <- t(as.matrix(iris[,1:4]))
y <- as.integer(iris[,5]) - 1   # mlpack prefers {0, 1, 2}
decisionTree(X, y)
}
