% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hoeffdingTrees}
\alias{hoeffdingTrees}
\title{Run a Random Forest classificatio}
\usage{
hoeffdingTrees(dataset, labels, pct = 0.3, nclasses = 7L)
}
\arguments{
\item{dataset}{A matrix of explanatory variables, i.e. \dQuote{features}}

\item{labels}{A vector of the dependent variable as integer values, i.e. \dQuote{labels}}

\item{pct}{A numeric value for the percentage of data to be retained for the test set}

\item{nclasses}{An integer value for the number of a distinct values in \code{labels}}
}
\value{
A list object
}
\description{
Run a Hoeffding Trees (Batch) Classifier
}
\details{
This function performs a Hoeffding Trees classification.
}
\examples{
data(covertype_small)                           # see help(covertype_small)
res <- hoeffdingTrees(covertype_small[-55,],    # features (already transposed)
                      covertype_small[55,],     # labels now in [0, 6] range
                      0.3)                      # percentage used for testing
str(res)  # accuracy varies as method is randomized but no seed set here
}
\seealso{
covertype_small
}
