% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loanData}
\alias{loanData}
\title{Loan data subset used for default prediction}
\source{
\url{https://datasets.mlpack.org/LoanDefault.csv}
}
\description{
A four column data set containing a binary variable \sQuote{Employed} (with zero denoting
unemployment and one employment), a numeric variable \sQuote{Bank Balance}, a numeric
variable \sQuote{Annual Salary} and a binary target variable \sQuote{Defaulted?} (with
zero denoting loan repayment and one denoting default).
}
\details{
The original source of the data is not documented by mlpack.
}
\references{
\url{https://archive.ics.uci.edu/dataset/31/covertype}
}
\keyword{data}
