% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logisticRegressionData}
\alias{logisticRegressionData}
\title{Logistic regression example data set}
\description{
A three column (synthetic) data set to illustrate logistic regression.
}
\keyword{data}
