\name{rcppmlpackexample-package}
\alias{rcppmlpackexample-package}
\alias{rcppmlpackexample}
\docType{package}
\title{\packageTitle{rcppmlpackexample}}
\description{\packageDescription{rcppmlpackexample}}
\section{Package Content}{\packageIndices{rcppmlpackexample}}
\author{\packageAuthor{rcppmlpackexample}}
\section{Maintainer}{\packageMaintainer{rcppmlpackexample}}
\keyword{package}
