\name{rcppmlpackexamples-package}
\alias{rcppmlpackexamples-package}
\alias{rcppmlpackexamples}
\docType{package}
\title{\packageTitle{rcppmlpackexamples}}
\description{\packageDescription{rcppmlpackexamples}}
\section{Package Content}{\packageIndices{rcppmlpackexamples}}
\author{\packageAuthor{rcppmlpackexamples}}
\section{Maintainer}{\packageMaintainer{rcppmlpackexamples}}
\keyword{package}
