% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logisticRegression}
\alias{logisticRegression}
\title{Run logistic regression}
\usage{
logisticRegression(data, labels, lambda = 0)
}
\arguments{
\item{data}{A matrix of data values}

\item{labels}{A vector of class labels}

\item{lambda}{An optional L2 regularization parameter, defaults to zero}
}
\value{
A list with predictions, probabilities and parmeters
}
\description{
Run a logistic regression returning classification
}
\details{
This function performs a logistic regression on the given data set.
}
\examples{
data(logisticRegression)
X <- as.matrix(logisticsData[, 1:2])
y <- as.matrix(logisticsData[, 3])
res <- logisticRegression(X, y)
res$parameters
}
